Fix page overflow not being made visible on render start
The UI page's overflow is hidden while Wray initializes, to hide the
fact that the rest of the UI is positioned just outside of the page's
boundaries, waiting to be slid into view when rendering begins. Once
the content has been slid into view, the overflow doesn't need to
be hidden anymore, since only valid overflow remains.